+++ aggreg_operators/maxaggregoperator.cpp
@@ -4,7 +4,7 @@
 
 #include <QDebug>
 
-int MaxAggregOperator::GetWorthyValue(const QList<long double>& t_list)
+int MaxAggregOperator::GetWorthyValue(const QList<double>& t_list)
 {
     if ( true == t_list.isEmpty() )
     {
