+++ aggreg_operators/medaggregoperator.h
@@ -19,7 +19,7 @@ private:
     // == METHODS ==
 public:
     MedAggregOperator();
-    int GetWorthyValue(const QList<long double> &t_list);
+    int GetWorthyValue(const QList<double>& t_list);
 
 protected:
     void ResetValues();
